lab1/app/requirements.txt
pandas
psycopg2-binary
regex
memory-profiler